Working on HitTest, fixed a value in DeclarationNumber

+++ GameLib/Item.cpp
@@ -70,7 +70,7 @@ void Item::Draw(std::shared_ptr<wxGraphicsContext> graphics)
  * @param y Y position to test
  * @return true if hit.
  */
-bool Item::HitTest(int x, int y)
+bool Item::HitTest(double x, double y)
 {
     double wid = mItemBitmap->GetWidth();
     double hit = mItemBitmap->GetHeight();
